Restructuring Containers for better connections
