--- src/containers/save/recommendations/header/header.scss
@@ -1,24 +1,0 @@
-@import '../../../../Common/_styles/colors';
-
-.header {
-  color: $darksmoke;
-  font-size: 12px;
-  font-weight: 400;
-  letter-spacing: 0.1px;
-  line-height: 16px;
-  margin: 0;
-  padding: 0.5em 0 0.8em;
-  text-align: center;
-  text-transform: uppercase;
-
-  a {
-    color: $teal;
-    display: block;
-    text-align: center;
-
-    &:hover {
-      color: darken($teal, 10%);
-    }
-  }
-}
-
