--- src/containers/save/recommendations/list/item.scss
@@ -1,195 +1,0 @@
-@import '../../../../Common/_styles/colors';
-@import '../../../../Common/_styles/variables';
-@import '../../../../Common/_styles/components';
-
-.list {
-  all: unset;
-  display: block;
-  font-family: $fontstack-default;
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  text-align: left;
-}
-
-.item {
-  background-color: $powder;
-  border-top: 1px solid $smoke;
-
-  &:hover {
-    background-color: $white;
-
-    .image {
-      border-color: $white;
-    }
-  }
-}
-
-.itemContainer,
-.spocContainer {
-  box-sizing: border-box;
-  display: block;
-  height: 110px;
-  margin: 0;
-  overflow: hidden;
-  padding: 10px;
-  position: relative;
-  width: 100%;
-
-  .hasImage & {
-    padding-right: 92px;
-  }
-}
-
-.spocContainer {
-  height: 131px;
-  padding: 8px 10px 10px;
-}
-
-.title {
-  box-sizing: border-box;
-  color: $pitch;
-  display: block;
-  font-size: 13px;
-  font-weight: 300;
-  height: 3.75em;
-  line-height: 1.25em;
-  max-width: 296px;
-  overflow: hidden;
-  padding: 0 0 10px;
-  text-align: left;
-}
-
-.link {
-  cursor: pointer;
-  display: block;
-  overflow: hidden;
-  text-decoration: none;
-  text-overflow: ellipsis;
-
-  &:link,
-  &:visited,
-  &:active {
-    color: $black;
-  }
-
-  &:hover {
-    color: $teal;
-  }
-}
-
-.image {
-  background-color: $smoke;
-  border: 5px solid $powder;
-  border-radius: 8px;
-  box-sizing: border-box;
-  height: 108px;
-  position: absolute;
-  right: 0;
-  top: 0;
-  width: 90px;
-  z-index: 2;
-
-  .isSpoc & {
-    height: 129px;
-  }
-}
-
-.source {
-  color: $silver;
-  font-size: 10px;
-  line-height: 1em;
-  margin: 3px 0 9px;
-}
-
-.actions {
-  display: block;
-}
-
-.save {
-  @include pocketButton;
-
-  background-color: transparent;
-  border: 0;
-  color: $overcast;
-  cursor: pointer;
-  display: block;
-  font-size: 12px;
-  font-weight: 100;
-  line-height: 16px;
-  text-decoration: none;
-
-  &:hover {
-    border: 0;
-    color: $hot-coral;
-  }
-}
-
-.sponsorHeader {
-  display: flex;
-  font-size: 12px;
-  justify-content: space-between;
-  align-items: baseline;
-  line-height: 20px;
-  padding: 8px 0px 2px 8px;
-  position: relative;
-}
-
-.sponsorAvatar {
-  border-radius: 50%;
-  display: inline-block;
-  height: 20px;
-  margin-right: 5px;
-  vertical-align: top;
-  width: 20px;
-}
-
-.sponsorLink {
-  color: $pitch;
-  font-weight: 600;
-  text-decoration: none;
-}
-
-.sponsorMenu {
-  float: right;
-}
-
-.sponsorLabel {
-  color: $overcast;
-  cursor: pointer;
-  font-size: 10px;
-  font-weight: 200;
-  line-height: inherit;
-  text-transform: uppercase;
-  vertical-align: baseline;
-}
-
-.sponsorDropdown {
-  background: $white;
-  border-radius: 4px;
-  box-shadow: 0 3px 8px 1px rgba(0, 0, 0, 0.5);
-  list-style-type: none;
-  margin: 0;
-  padding: 7px 0;
-  position: absolute;
-  right: 10px;
-  top: 25px;
-  z-index: 3;
-
-  button,
-  a {
-    @include pocketButton;
-
-    box-sizing: border-box;
-    color: $pitch;
-    display: block;
-    font-size: 16px;
-    padding: 4px 10px;
-    width: 100%;
-
-    &:hover {
-      background: $teal;
-      color: $white;
-    }
-  }
-}
