--- src/containers/save/save.html
@@ -1,18 +1,0 @@
-<link rel="stylesheet" href="/fonts/fonts.css">
-<style>
-  html, body  {
-      font-family:
-        'proxima-nova',
-        'HelveticaNeue',
-        'Helvetica Neue',
-        Helvetica,
-        Arial,
-        sans-serif;
-    background:none transparent;
-    margin: 0;
-    border: 0;
-    padding: 0;
-    overflow: hidden;
-  }
-</style>
-<div id="pocket-extension-root"></div>
