--- src/containers/save/save.scss
@@ -1,14 +1,0 @@
-@import '../../Common/_styles/colors';
-@import '../../Common/_styles/variables';
-@import '../../Common/_styles/components';
-
-.hanger {
-  @include fontBase;
-
-  font-size: 16px;
-  position: fixed;
-  right: 10px;
-  top: 15px;
-  width: 320px;
-  z-index: 2147483647;
-}
