--- src/containers/save/toolbar/tagging/chips/chips.scss
@@ -1,60 +1,0 @@
-@import '../../../../../Common/_styles/colors';
-
-.chipsList {
-  display: inline;
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-
-  &:last-child {
-    min-height: 21px;
-  }
-}
-
-.chip {
-  background: $powder;
-  border: 1px solid $gray;
-  border-radius: 3px;
-  color: $pitch;
-  cursor: pointer;
-  display: inline-block;
-  font-weight: 300;
-  line-height: 16px;
-  margin-bottom: 2px;
-  margin-right: 3px;
-  margin-top: 2px;
-  padding: 2px 12px;
-  text-transform: lowercase;
-
-  &:first-child {
-    margin-left: 20px;
-  }
-
-  span {
-    display: inline-block;
-    font-size: 18px;
-    font-weight: 100;
-    line-height: 15px;
-    padding-left: 4px;
-    vertical-align: middle;
-
-    &:hover {
-      color: $hot-coral;
-    }
-  }
-
-  &.active {
-    background-color: $smoke;
-    padding: 2px 5px 2px 6px;
-
-    // &:hover {
-    //     background-color: $smoke;
-    // }
-  }
-
-  &:hover {
-    // background-color: $gray;
-    border-color: $overcast;
-  }
-}
-
