--- src/containers/save/toolbar/tagging/tagging.scss
@@ -1,72 +1,0 @@
-@import '../../../../Common/_styles/colors';
-@import '../../../../Common/_styles/variables';
-
-.tagging {
-  font-family: $fontstack-default;
-  padding: 5px 0 0;
-  position: relative;
-}
-
-.well {
-  background: $white;
-  background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" fill="#999999"><path d="M83.38 54.6L43.8 17.02H17.41V43.4l39.58 39.58L83.38 54.6zM30.95 36.81c-3.45 0-6.25-2.67-6.25-5.96s2.8-5.97 6.25-5.97 6.25 2.67 6.25 5.97c0 3.29-2.8 5.96-6.25 5.96z"/></svg>');
-  background-position: 4px 7px;
-  background-repeat: no-repeat;
-  background-size: 22px;
-  border: 1px solid $smoke;
-  border-radius: 3px;
-  box-sizing: border-box;
-  font-size: 13px;
-  line-height: 16px;
-  margin: 0;
-  padding: 4px 10px;
-  position: relative;
-  text-align: left;
-
-  &.active {
-  }
-}
-
-.placeholder {
-  color: $overcast;
-  left: 29px;
-  // pointer-events: none;
-  position: absolute;
-  top: 10px;
-}
-
-.typeaheadWrapper {
-  position: relative;
-}
-
-.typeaheadList {
-  background: $white;
-  border: 1px solid $smoke;
-  border-radius: 0 0 3px 3px;
-  border-top: none;
-  box-shadow: 0 4px 4px rgba(0, 0, 0, 0.2);
-  box-sizing: border-box;
-  display: block;
-  left: 0;
-  list-style-type: none;
-  margin: 0;
-  max-height: 8.8em;
-  overflow-x: hidden;
-  overflow-y: auto;
-  padding: 2px 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-}
-
-.typeahead {
-  cursor: pointer;
-  display: block;
-  padding: 2px 8px;
-
-  &:hover,
-  &.active {
-    background-color: $teal;
-    color: $white;
-  }
-}
