--- src/containers/save/toolbar/toolbar.scss
@@ -1,37 +1,0 @@
-@import '../../../Common/_styles/colors';
-@import '../../../Common/_styles/variables';
-@import '../../../Common/_styles/components';
-
-.toolbar {
-  @include pocketPanel;
-
-  box-shadow: $panel-shadow;
-}
-
-.toolbar {
-  color: $pitch;
-  font-family: $fontstack-default;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 1.5;
-  padding: 8px 10px;
-  position: relative;
-  text-align: left;
-  z-index: 100;
-}
-
-.statusText {
-  display: inline-block;
-  font-weight: 600;
-  padding: 3px 0;
-}
-
-.error {
-  background-color: $snow;
-  border-radius: 4px;
-  box-shadow: inset 0 0 4px rgba(0, 0, 0, 0.5);
-  color: $black;
-  font-size: 14px;
-  margin-top: 10px;
-  padding: 10px;
-}
