--- src/containers/sites/twitter/twitter.scss
@@ -1,12 +1,0 @@
-.pocketIcon {
-  display: inline-block;
-  fill: currentColor;
-  height: 16px;
-  vertical-align: bottom;
-  width: 16px;
-
-  &:hover,
-  .saved & {
-    fill: #ef4056;
-  }
-}
